Legacy cleanup, part 2 (#3659)
* Legacy cleanup, part 2

There are still things that can be removed, those will be likely done in part three:
`The Return of the Cleanup` :)


* Fix windows build

Somehow there were some unfixed errors from enabled warnings?
I'm unsure if I've pulled repo in unstable state, or those were somehow
missed.

+++ Code/Legacy/CryCommon/ILocalizationManager.h
@@ -73,11 +73,11 @@ struct SLocalizedSoundInfoGame
     bool    bIsIntercepted;
 
     // SoundMoods.
-    int         nNumSoundMoods;
+    size_t         nNumSoundMoods;
     SLocalizedAdvancesSoundEntry* pSoundMoods;
 
     // EventParameters.
-    int         nNumEventParameters;
+    size_t         nNumEventParameters;
     SLocalizedAdvancesSoundEntry* pEventParameters;
 };
